Movie & Banner: Fixed and improved movie and banner endpoints
Closes #34+

+++ src/movies/movie.controller.ts
@@ -1,73 +1,33 @@
 import {
-  Controller,
-  Get,
-  Post,
   Body,
-  Param,
-  Query,
-  Patch,
+  Controller,
   Delete,
-  UseInterceptors,
+  Get,
+  Param,
+  Patch,
+  Post,
+  Query,
+  UploadedFile,
   UploadedFiles,
+  UseInterceptors,
 } from '@nestjs/common';
-import createSingleMulterStorage from '../config/multerStorage';
-import { MovieService } from './movie.service';
-import { CreateMovieDto } from './dto/create.movie.dto';
-import { UpdateMovieDto } from './dto/update.movie.dto';
 import { FileFieldsInterceptor } from '@nestjs/platform-express';
-import {
-  ApiTags,
-  ApiOperation,
-  ApiResponse,
-  ApiQuery,
-  ApiParam,
-  ApiConsumes,
-  ApiBody,
-} from '@nestjs/swagger';
+import { ApiConsumes, ApiExtraModels, ApiResponse } from '@nestjs/swagger';
+import { StatusCodes } from 'http-status-codes';
+import { join } from 'path';
+import { ServiceResponse } from '../common/model/service-response';
+import { createSingleMulterStorage } from '../config/multerStorage';
+import { CreateMovieDto } from './dto/create.movie.dto';
+import { MovieQueries } from './dto/movie-queries.dto';
+import { UpdateMovieDto } from './dto/update.movie.dto';
+import { MovieService } from './movie.service';
 
-@ApiTags('Movies')
 @Controller('movies')
 export class MovieController {
   constructor(private readonly movieService: MovieService) {}
 
-  @Post('')
-  @ApiOperation({ summary: 'Create a new movie' })
+  @Post()
   @ApiConsumes('multipart/form-data')
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        title: { type: 'string' },
-        descriptions: { type: 'string', nullable: true },
-        author: { type: 'string', nullable: true },
-        image: {
-          type: 'string',
-          format: 'binary',
-          nullable: true,
-        },
-        trailer: {
-          type: 'string',
-          format: 'binary',
-          nullable: true,
-        },
-        type: {
-          type: 'string',
-          enum: ['2D', '3D'],
-        },
-        duration_min: {
-          type: 'integer',
-          description: 'Duration of the movie in minutes',
-          example: 120,
-        },
-        release_date: {
-          type: 'string',
-          format: 'date-time',
-        },
-      },
-    },
-  })
-  @ApiResponse({ status: 201, description: 'Movie created successfully' })
-  @ApiResponse({ status: 400, description: 'Bad Request' })
   @UseInterceptors(
     FileFieldsInterceptor(
       [
@@ -77,85 +37,47 @@ export class MovieController {
       { storage: createSingleMulterStorage(true, true) },
     ),
   )
+  @ApiResponse({ type: ServiceResponse })
   async create(
     @UploadedFiles()
-    files: { image?: Express.Multer.File[]; trailer?: Express.Multer.File[] },
-    @Body() createMovieDto: CreateMovieDto,
+    files: { image: Express.Multer.File[]; trailer: Express.Multer.File[] },
+    @Body() dto: CreateMovieDto,
   ) {
-    if (files.image) {
-      createMovieDto.image = files.image
-        .map((file) => `uploads/movies/image/${file.filename}`)
-        .join(',');
-    }
-    if (files.trailer && files.trailer[0]) {
-      createMovieDto.trailer = `uploads/movies/trailer/${files.trailer[0].filename}`;
-    }
+    if (files.image) dto.image = `uploads/${files.image[0].filename}`;
+    if (files.trailer) dto.trailer = `uploads/${files.trailer[0].filename}`;
 
-    console.log(createMovieDto);
-    return this.movieService.create(createMovieDto);
+    const newMovie = await this.movieService.create(dto);
+    return ServiceResponse.success(
+      'Movie added successfully',
+      { id: newMovie.id },
+      StatusCodes.CREATED,
+    );
   }
 
   @Get()
-  @ApiOperation({ summary: 'Get all movies' })
-  @ApiQuery({ name: 'page', required: false, type: Number })
-  @ApiQuery({ name: 'limit', required: false, type: Number })
-  @ApiQuery({ name: 'sort', required: false, type: String })
-  @ApiQuery({ name: 'order', required: false, type: String })
-  @ApiQuery({ name: 'search', required: false, type: String })
-  @ApiResponse({ status: 200, description: 'List of movies' })
-  async findAll(@Query() query: any) {
-    return this.movieService.findAll(query);
+  @ApiExtraModels(MovieQueries)
+  @ApiResponse({ type: ServiceResponse })
+  async findAll(@Query() query: MovieQueries) {
+    const movies = await this.movieService.findAll(query);
+    return ServiceResponse.success('Got all movies', movies);
   }
 
   @Get(':id')
-  @ApiOperation({ summary: 'Get a movie by ID' })
-  @ApiParam({ name: 'id', required: true, type: Number })
-  @ApiResponse({ status: 200, description: 'Movie found' })
-  @ApiResponse({ status: 404, description: 'Movie not found' })
-  @ApiResponse({ status: 400, description: 'Bad Request' })
+  @ApiResponse({ type: ServiceResponse })
   async findOne(@Param('id') id: number) {
-    return this.movieService.findOne(id);
+    const movie = await this.movieService.findOne(id);
+    if (!movie) {
+      return ServiceResponse.failure(
+        `Movie #${id} not found`,
+        null,
+        StatusCodes.NOT_FOUND,
+      );
+    }
+    return ServiceResponse.success(`Found movie #${id}`, movie);
   }
 
   @Patch(':id')
-  @ApiOperation({ summary: 'Update a movie by ID' })
   @ApiConsumes('multipart/form-data')
-  @ApiParam({ name: 'id', required: true, type: Number })
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        title: { type: 'string' },
-        descriptions: { type: 'string', nullable: true },
-        author: { type: 'string', nullable: true },
-        image: {
-          type: 'string',
-          format: 'binary',
-          nullable: true,
-        },
-        trailer: {
-          type: 'string',
-          format: 'binary',
-          nullable: true,
-        },
-        type: {
-          type: 'string',
-          enum: ['2D', '3D'],
-        },
-        duration_min: {
-          type: 'integer',
-          description: 'Duration of the movie in minutes',
-        },
-        release_date: {
-          type: 'string',
-          format: 'date-time',
-        },
-      },
-    },
-  })
-  @ApiResponse({ status: 200, description: 'Movie updated successfully' })
-  @ApiResponse({ status: 404, description: 'Movie not found' })
-  @ApiResponse({ status: 400, description: 'Bad Request' })
   @UseInterceptors(
     FileFieldsInterceptor(
       [
@@ -165,29 +87,31 @@ export class MovieController {
       { storage: createSingleMulterStorage(true, true) },
     ),
   )
+  @ApiResponse({ type: ServiceResponse })
   async update(
     @Param('id') id: number,
     @UploadedFiles()
-    files: { image?: Express.Multer.File[]; trailer?: Express.Multer.File[] },
-    @Body() updateMovieDto: UpdateMovieDto,
+    files: { image: Express.Multer.File[]; trailer: Express.Multer.File[] },
+    @Body() dto: UpdateMovieDto,
   ) {
-    if (files.image) {
-      updateMovieDto.image = files.image.map((file) => file.filename).join(',');
-    }
-    if (files.trailer && files.trailer[0]) {
-      updateMovieDto.trailer = files.trailer[0].filename;
-    }
+    if (files.image) dto.image = `uploads/${files.image[0].filename}`;
+    if (files.trailer) dto.trailer = `uploads/${files.trailer[0].filename}`;
 
-    return this.movieService.update(id, updateMovieDto);
+    await this.movieService.update(id, dto);
+    return ServiceResponse.success(`Updated movie #${id}`, null);
   }
 
   @Delete(':id')
-  @ApiOperation({ summary: 'Delete a movie by ID' })
-  @ApiParam({ name: 'id', required: true, type: Number })
-  @ApiResponse({ status: 200, description: 'Movie deleted successfully' })
-  @ApiResponse({ status: 404, description: 'Movie not found' })
-  @ApiResponse({ status: 400, description: 'Bad Request' })
+  @ApiResponse({ type: ServiceResponse })
   async remove(@Param('id') id: number) {
-    return this.movieService.remove(id);
+    if (!(await this.findOne(id)).success) {
+      return ServiceResponse.failure(
+        `Movie #${id} not found`,
+        null,
+        StatusCodes.NOT_FOUND,
+      );
+    }
+    await this.movieService.remove(id);
+    return ServiceResponse.success(`Deleted movie #${id}`, null);
   }
 }
